% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{as.tbl_genome}
\alias{as.tbl_genome}
\alias{as.tbl_genome.tbl_df}
\alias{as.tbl_genome.data.frame}
\title{Coerce objects to tbl_genome.}
\usage{
as.tbl_genome(x)

\method{as.tbl_genome}{tbl_df}(x)

\method{as.tbl_genome}{data.frame}(x)
}
\arguments{
\item{x}{object to convert to tbl_genome.}
}
\value{
\code{\link[=tbl_genome]{tbl_genome()}}
}
\description{
This is an S3 generic. valr includes methods to coerce tbl_df and data.frame
objects.
}
