% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.r
\name{bam_to_df}
\alias{bam_to_df}
\title{Read in bam file as a trbl_interval}
\usage{
bam_to_df(filename = NULL)
}
\arguments{
\item{filename}{path to bam file}
}
\description{
Read in bam file as a trbl_interval
}
