% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_closest.r
\name{bed_closest}
\alias{bed_closest}
\title{Identify closest intervals.}
\usage{
bed_closest(x, y, overlap)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{overlap}{include overlapping intervals}
}
\value{
\code{data_frame}
}
\description{
Identify closest intervals.
}
\examples{

}
\seealso{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/closest.html}
}

