% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_cluster.r
\name{bed_cluster}
\alias{bed_cluster}
\title{Cluster neighboring intervals.}
\usage{
bed_cluster(bed_tbl, max_dist = 0, strand = FALSE)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{max_dist}{maximum distance between clustered intervals.
default is 0 to cluster book-ended intervals.}

\item{strand}{cluster features on same strand}
}
\value{
\code{data_frame}
}
\description{
returned \code{data_frame} contains a new \code{.id} column that can be used for
grouping along with \code{chrom}.
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100,  200,
 "chr1", 180,  250,
 "chr1", 501,  1000
)

bed_cluster(bed_tbl)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/cluster.html}
}

