% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_complement.r
\name{bed_complement}
\alias{bed_complement}
\title{identify intervals in a genome that are not covered by a query}
\usage{
bed_complement(bed_df, genome)
}
\arguments{
\item{bed_df}{BED data.frame}

\item{genome}{chrom sizes}
}
\value{
\code{data.frame}
}
\description{
identify intervals in a genome that are not covered by a query
}
\examples{

bed_df <- dplyr::tibble(
   ~chrom, ~start, ~end,
   "chr1", 100,    300,
   "chr2", 200,    400,
   "chr3", 500,    600
)

bed_complement(bed_df)

}

