% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_flank.r
\name{bed_flank}
\alias{bed_flank}
\title{flank}
\usage{
bed_flank(bed_df, size = 0, both = FALSE, left = FALSE, right = FALSE,
  strand = FALSE, fraction = 0)
}
\arguments{
\item{left}{number of bases on left side}

\item{right}{number of bases on right side}

\item{strand}{define -l and -r based on strand}

\item{percent}{define flanks based on fraction of feature's length}
}
\value{
\code{dplyr::tbl_df}
}
\description{
make flanks from input BED regions
}

