% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_jaccard.r
\name{bed_jaccard}
\alias{bed_jaccard}
\title{Calculate jaccard statistics on two sets of intervals.}
\usage{
bed_jaccard(x, y)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}
}
\value{
\code{data_frame} with the following columns:
 \itemize{
   \item{\code{len_i}}{ length of the intersection}
   \item{\code{len_u}}{ length of the union}
   \item{\code{jaccard}}{ jaccard statistic}
   \item{\code{n_int}}{ number of intersecting intervals between x and y}
 }
}
\description{
Calculate jaccard statistics on two sets of intervals.
}
\examples{
x <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 10,     20,
  "chr1", 30,     40
)
 
y <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", 15,     20
)
 
bed_jaccard(x, y)
 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/jaccard.html}
}

