% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_makewindows.r
\name{bed_makewindows}
\alias{bed_makewindows}
\title{bed_makewindows}
\usage{
bed_makewindows(bed_df, genome, win_size = 0, step_size = 0,
  num_windows = 0, reverse = FALSE)
}
\arguments{
\item{bed_df}{BED data in \code{dplyr::tbl_df} format}

\item{genome}{genome file with chromosome sizes}

\item{win_size}{divide intervals into fixed-size windows}

\item{step_size}{size to step before next window}

\item{reverse}{reverse window numbers?}

\item{num_wins}{divide intervals to fixed number of windows}

\item{win_names}{one of 'name', 'num', 'namenum'}
}
\value{
\code{dplyr::tbl_df} with `win.num` column
}
\description{
Divide intervals into new intervals ("windows") with labels
}
\examples{

genome <- dplyr::tibble(
 ~chrom, ~size,
 "chr1", 5000,
 "chr2", 400
)

 bed_df <- dplyr::tibble(
   ~chrom, ~start, ~end, ~name, ~score, ~strand,
   "chr1", 100, 200, 'A', '.', '+',
   "chr2", 300, 350, 'B', '.', '-'
 ) 
 
}

