% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_merge.r
\name{bed_merge}
\alias{bed_merge}
\title{Merge overlapping intervals}
\usage{
bed_merge(bed_tbl)
}
\description{
Merge overlapping intervals
}
\examples{
bed_tbl <- dplyr::tibble(
 ~chrom, ~start, ~end,
 "chr1", 100,    200,
 "chr1", 150,    250,
 "chr2", 200,    400
)

bed_tbl <- bed_sort(bed_tbl) 
bed_merge(bed_tbl)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/merge.html}
}

