% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_projection.r
\name{bed_projection}
\alias{bed_projection}
\title{Projection Test for query interval overlap.}
\usage{
bed_projection(x, y, genome, by_chrom = FALSE)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{genome}{chrom sizes}

\item{by_chrom}{compute test per chromosome  rather than across whole genome}
}
\value{
\code{data_frame} with the following columns:
  \itemize{ 
    \item{\code{chrom}} {the name of chromosome tested if \code{by_chrom} is \code{TRUE}, otherwise set to \code{whole_genome}}
    \item{\code{.p_value}} {p-value from a binomial test, note that p-values > 0.5 will be reported as 1 - p-value and .lower_tail will be set to FALSE}
    \item{\code{.effect_ratio}} {ratio of observed overlap probabilities to expected}
    \item{\code{.lower_tail}} {TRUE denotes thate observed probabilities are in the lower tail of the distribution (less overlap than expected), FALSE
    denotes that the observed probabililty are in the upper tail of the distribution (more overlap than expected)}
    }
}
\description{
Projection Test for query interval overlap.
}
\details{
input tbls can be grouped using \code{\link[dplyr]{group_by}} prior to 
  analysis. Input tbls are grouped by \code{chrom} by default. Grouping by 
  \code{strand} will constrain analyses to the same strand. To compare 
  opposing strands between two tbls, strands on the \code{y} tbl can first be
  be inverted using \code{\link{flip_strands}}.
}
\examples{
genome <- tibble::tribble(
 ~chrom, ~size,
 "chr1", 1e6,
 "chr2", 2e6,
 "chr3", 4e6
)

x <- bed_random(genome)
bed_shuffle(x, genome)

}
\seealso{
\url{http://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1002529}

Other interval-stats: \code{\link{bed_absdist}},
  \code{\link{bed_fisher}}, \code{\link{bed_jaccard}},
  \code{\link{bed_reldist}}
}

