% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_shuffle.r
\name{bed_shuffle}
\alias{bed_shuffle}
\title{shuffle input intervals}
\usage{
bed_shuffle(x, genome, incl = NULL, excl = NULL, max_tries = 100)
}
\arguments{
\item{x}{tbl of intervals}

\item{genome}{chrom sizes}

\item{incl}{tbl of included intervals}

\item{excl}{tbl of excluded intervals}

\item{max_tries}{maximum tries to identify a bounded region}
}
\value{
\code{data_frame}
}
\description{
shuffle input intervals
}
\examples{
x <- tibble::frame_data(
)

genome <- tibble::frame_data(
)

}
\seealso{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/shuffle.html}
}

