% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_sort.r
\name{bed_sort}
\alias{bed_sort}
\title{Sort intervals}
\usage{
bed_sort(intervals)
}
\arguments{
\item{intervals}{tbl of intervals}
}
\description{
Sort intervals
}
\examples{
bed_df <- dplyr::tibble(
   ~chrom, ~start, ~end,
   "chr8", 500, 1000,
   "chr8", 1000, 5000,
   "chr1", 100, 300,
   "chr1", 100, 200
)

bed_df \%>\% bed_sort()
}

