% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_window.r
\name{bed_window}
\alias{bed_window}
\title{Identify intersecting intervals within a specified distance}
\usage{
bed_window(x, y, w = 1000, l = 0, r = 0, strand = FALSE,
  strand_opp = FALSE)
}
\arguments{
\item{x}{BED intervals}

\item{y}{BED intervals}

\item{w}{basepairs upstream and downstream to add to x features}

\item{strand}{intersect intervals on same strand}
}
\description{
Identify intersecting intervals within a specified distance
}
\examples{
x <- tibble::frame_data(
~chrom, ~start, ~end,
"chr1", 10,    100,
"chr2", 200,    400,
"chr2", 300,    500,
"chr2", 800,    900
)

y <- tibble::frame_data(
~chrom, ~start, ~end,
"chr1", 150,    400,
"chr2", 230,    430,
"chr2", 350,    430
)

bed_intersect(x, y)
bed_window(x, y, w = 100)

 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/window.html}
}

