% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makewindows.r
\name{bedtools_makewindows}
\alias{bedtools_makewindows}
\title{makewindows.R}
\usage{
bedtools_makewindows(bed_df, reverse = FALSE)
}
\arguments{
\item{bed_df}{BED data in \code{dplyr::tbl_df} format}

\item{reverse}{reverse window numbers}
}
\value{
\code{dplyr::tbl_df} with `win.num` column
}
\description{
dplyr implementation of BEDtools makewindows
}

