% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{bedtools_map}
\alias{bedtools_map}
\alias{bt_map}
\title{map.r}
\usage{
bedtools_map(a, b)

bt_map(a, b)
}
\arguments{
\item{a}{intervals in tbl_df format}

\item{b}{signal in tbl_df format}

\item{col_name}{bare column name}

\item{operation}{operation to perform on intersected intervals. One of mean, median,
sum, min, max, absmin, absmax}
}
\value{
\code{dplyr::tbl_df}
}
\description{
impelementation of BEDtools `map` function with dplyr
}

