% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome.r
\name{bound_intervals}
\alias{bound_intervals}
\title{select intervals bounded by a genome}
\usage{
bound_intervals(bed_tbl, genome, trim = FALSE)
}
\arguments{
\item{bed_tbl}{a tbl of intervals}

\item{genome}{a tbl of chrom sizes}

\item{trim}{adjust coordinates for out-of-bounds intervals}
}
\value{
\code{data_frame}
}
\description{
select intervals bounded by a genome
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", -100,   500,
 "chr1", 100,    1e9,
 "chr1", 500,    1000
)

genome <- read_genome('https://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes')

bound_intervals(bed_tbl, genome)
bound_intervals(bed_tbl, genome, trim = TRUE)

}

