% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.r
\name{create_introns}
\alias{create_introns}
\title{Create intron features.}
\usage{
create_introns(x)
}
\arguments{
\item{x}{\link{tbl_interval} in BED12 format}
}
\description{
Numbers in the \code{score} column are intron numbers from 5' to 3' independent of
strand. I.e., the first introns for \code{+} and \code{-} strand genes both have \code{score}
values of \code{1}.
}
\examples{
x <- read_bed12(valr_example('mm9.refGene.bed.gz'))

create_introns(x)

}
