% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{format_bed}
\alias{format_bed}
\title{reformat bed tbl to match another tbl}
\usage{
format_bed(x, y)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}
}
\description{
\code{format_bed} returns a tbl whose columns are ordered by another tbl.
Columns not found in y tbl are dropped. Missing y columns are added and populated
with a dummy entry "."
}

