% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-external.r
\name{id_var}
\alias{id_var}
\title{Plyr function id_var packaged due to plyr being retired
Numeric id for a vector.}
\usage{
id_var(x, drop = FALSE)
}
\description{
Plyr function id_var packaged due to plyr being retired
Numeric id for a vector.
}
\keyword{internal}
