% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{is.tbl_genome}
\alias{is.tbl_genome}
\title{Test if the object is a tbl_genome.}
\usage{
is.tbl_genome(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{\link[=tbl_genome]{tbl_genome()}} class.
}
\description{
Test if the object is a tbl_genome.
}
