% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{is.tbl_interval}
\alias{is.tbl_interval}
\title{Test if the object is a tbl_interval.}
\usage{
is.tbl_interval(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{\link[=tbl_interval]{tbl_interval()}} class.
}
\description{
Test if the object is a tbl_interval.
}
