% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{is.tbl_sizes}
\alias{is.tbl_sizes}
\title{Test if the object is a tbl_sizes.}
\usage{
is.tbl_sizes(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{tbl_sizes} class.
}
\description{
Test if the object is a tbl_sizes.
}
