% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_impl}
\alias{random_impl}
\title{Generate random intervals on a genome.}
\usage{
random_impl(genome, length, n, seed = 0L)
}
\arguments{
\item{genome}{tbl of genome sizes}

\item{length}{legnth of output intervals}

\item{n}{number of random intervals}

\item{seed}{integer seed for reproducible intervals}
}
\value{
\code{data_frame}
}
\description{
Generate random intervals on a genome.
}

