\name{Rbedtools-package}
\alias{Rbedtools-package}
\alias{Rbedtools}
\docType{package}
\title{
\packageTitle{Rbedtools}
}
\description{
\packageDescription{Rbedtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rbedtools}
\packageIndices{Rbedtools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rbedtools}

Maintainer: \packageMaintainer{Rbedtools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
