% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbedtools.r
\docType{package}
\name{Rbedtools}
\alias{Rbedtools}
\alias{Rbedtools-package}
\title{Rbedtools: genome arithmetic in R}
\description{
Rbedtools provides tools to read and manipulate intervals and signals on a 
genome reference. While other analysis suites like BEDtools and BEDOPS enable
fast command line analysis, Rbedtools enables similar analyses \emph{within
R}.
}
\details{
Rbedtools focuses on manipulating intervals in BED format and signal in 
bedGraph format. Eventually this package will power interactive 
visualizations of genome-scale data with \code{shiny}.

To learn more about Rbedtools, start with the vignette: 
\code{browseVignettes(package = "Rbedtools")}
}
\author{
Jay Hesselberth <jay.hesselberth@gmail.com>
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/index.html}

\url{https://pythonhosted.org/pybedtools/}

\url{http://bedops.readthedocs.org/en/latest/index.html}

\url{https://bioconductor.org/packages/release/bioc/html/GenomicRanges.html}
}

