% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_bed}
\alias{read_bed}
\alias{read_bedgraph}
\title{Methods to read BED and related files}
\usage{
read_bed(filename, n_fields = 3, col_names = bed12_colnames, sort = TRUE,
  ...)

read_bedgraph(filename)
}
\arguments{
\item{filename}{file or URL}

\item{n_fields}{number fields in the BED file}

\item{col_names}{column names to use for \code{dplyr::tbl_df}}

\item{sort}{sort the tbl by chrom and start}

\item{...}{options to pass to \code{readr::read_tsv}}

\item{filename}{file name or URL}
}
\value{
\code{dplyr::tbl_df}
}
\description{
Methods to read BED and related files
}

