% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_bigbed}
\alias{read_bigbed}
\title{Read a bigBed file into a valr compatible tbl}
\usage{
read_bigbed(path, ...)
}
\arguments{
\item{path}{path to bigBed file}

\item{...}{params for \code{cpp11bigwig::read_bigbed()}}
}
\description{
This function will output a tibble
zero-based chrom, start, end, value columns,
as well as any additional columns in the bigBed file.
}
\examples{
read_bigbed(valr_example("test.bb"))

}
