% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_bigwig}
\alias{read_bigwig}
\title{Read in a bigwig file into a valr compatible tbl}
\usage{
read_bigwig(path, set_strand = "+")
}
\arguments{
\item{path}{path to bigWig file}

\item{set_strand}{strand to add to output (defaults to "+")}
}
\description{
This function will output a 5 column tibble with
zero-based chrom, start, end, score, and strand columns.
}
\examples{

bw <- read_bigwig(valr_example('hg19.dnase1.bw'))
head(bw)

}
