% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_gtf}
\alias{read_gtf}
\title{Import and convert a GTF/GFF file into a valr compatible bed tbl format}
\usage{
read_gtf(path, ...)
}
\arguments{
\item{path}{path to gtf file}
}
\description{
This function will output a tibble with the
required chrom, start, and end columns, as well as other columns depending
on content in GTF/GFF file.
}
\details{
Based on the GTF specification at \url{https://useast.ensembl.org/info/website/upload/gff.html}
}
\examples{
gtf <- read_gtf(valr_example("hg19.gencode.gtf.gz"))
head(gtf)

}
