% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.r
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF file.}
\usage{
read_vcf(vcf)
}
\arguments{
\item{vcf}{vcf filename}
}
\value{
\code{data_frame}
}
\description{
Read a VCF file.
}
\note{
return value has \code{chrom}, \code{start} and \code{end} columns.
Interval lengths are the size of the 'REF' field.
}
\examples{
vcf_file <- valr_example('test.vcf.gz')
read_vcf(vcf_file)

}
\seealso{
Other read functions: \code{\link{read_bed}},
  \code{\link{read_genome}}
}
\concept{read functions}
