% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tibble.r
\name{reexports-deprecated}
\alias{reexports-deprecated}
\alias{data_frame}
\alias{as_data_frame}
\title{Deprecated tibble re-exports}
\usage{
data_frame(...)

as_data_frame(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link[tibble:tibble]{tibble::tibble()}} or \code{\link[tibble:as_tibble]{tibble::as_tibble()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{data_frame()} and \code{as_data_frame()} were re-exported from tibble, where they
have since been deprecated. Use \code{\link[tibble:tibble]{tibble::tibble()}} and \code{\link[tibble:as_tibble]{tibble::as_tibble()}}
instead.
}
\keyword{internal}
