% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-cpp11bigwig.r, R/reexport-tibble.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_bigwig}
\alias{read_bigbed}
\alias{tribble}
\alias{tibble}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cpp11bigwig}{\code{\link[cpp11bigwig:read_bigbed]{read_bigbed()}}, \code{\link[cpp11bigwig:read_bigwig]{read_bigwig()}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}, \code{\link[tibble:tibble]{tibble()}}, \code{\link[tibble:tribble]{tribble()}}}
}}

