% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{set_groups}
\alias{set_groups}
\title{Compare tbl_dfs and set y groups based on x tbl}
\usage{
set_groups(x, y)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}
}
\description{
\code{set_groups} return the x tbl grouped by the y tbls groups. 
Groups in X that are not found in the Y tbl are dropped from the grouping.
}

