% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{shared_groups}
\alias{shared_groups}
\title{Compare two tbl_dfs and find the minimum shared groups}
\usage{
shared_groups(x, y)
}
\arguments{
\item{x}{tbl}

\item{y}{tbl}
}
\value{
\code{list}
}
\description{
\code{shared_groups} returns the minimum shared groups between
\code{x} and \code{y} tbls. Return value is \code{NULL} if there are
no shared groups.
}
\examples{
x <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
  "chr1", 150,    400,  100,
  "chr2", 230,    430,  200,
  )

y <- tibble::frame_data(
  ~chrom,   ~start,    ~end,  ~value,
  "chr1",    50,       100,  1,
  "chr1",    100,       150,  2
  )
  
x <- dplyr::group_by(x, chrom, value)
y <- dplyr::group_by(y, chrom, value)
shared_groups(x, y)
 
y <- dplyr::group_by(y, chrom)
shared_groups(x, y)

y <-dplyr::ungroup(y)
shared_groups(x, y)
}

