% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{tbl_genome}
\alias{tbl_genome}
\alias{trbl_genome}
\title{Tibble for reference sizes.}
\usage{
tbl_genome(x, ..., .validate = TRUE)

trbl_genome(...)
}
\arguments{
\item{x}{A \code{data_frame}}

\item{...}{params for \code{\link[tibble:tibble]{tibble::tibble()}}}

\item{.validate}{check valid column names}
}
\value{
\code{\link[=tbl_genome]{tbl_genome()}}
}
\description{
Equivalent to information in UCSC "chromSizes" files. Required column names are:
\code{chrom} and \code{size}

Construct a tbl_genome using tribble formatting.
}
\examples{
genome <- tibble::tribble(
  ~chrom, ~size,
  'chr1', 1e6,
  'chr2', 1e7
)

is.tbl_genome(genome)
genome <- tbl_genome(genome)
is.tbl_genome(genome)

trbl_genome(
  ~chrom, ~size,
  'chr1', 1e6
)

}
