% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{tbl_interval}
\alias{tbl_interval}
\title{create an interval tibble}
\usage{
tbl_interval(...)
}
\arguments{
\item{...}{A set of name-value pairs. Arguments are evaluated sequentially,
so you can refer to previously created variables.}
}
\description{
Inherits from \code{[tibble](tibble)} and enforces \code{chrom}, \code{start}
and \code{end} columns.
}
