% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{trbl_genome}
\alias{trbl_genome}
\title{Construct a tbl_genome using tribble formatting.}
\usage{
trbl_genome(...)
}
\arguments{
\item{...}{for \code{\link[tibble:tribble]{tibble::tribble()}}}
}
\value{
\code{\link[=tbl_genome]{tbl_genome()}}
}
\description{
Construct a tbl_genome using tribble formatting.
}
\examples{
trbl_genome(
  ~chrom, ~size,
  'chr1', 1e6
)

}
