% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{trbl_interval}
\alias{trbl_interval}
\title{Construct an tbl_interval using tribble formatting.}
\usage{
trbl_interval(...)
}
\arguments{
\item{...}{data for \code{\link[tibble:tribble]{tibble::tribble()}}}
}
\value{
\code{\link[=tbl_interval]{tbl_interval()}}
}
\description{
Construct an tbl_interval using tribble formatting.
}
\examples{
trbl_interval(
  ~chrom, ~start, ~end,
  'chr1',  1,      50,
  'chr1',  10,     75
)

}
