% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valr-package.r
\docType{package}
\name{valr-package}
\alias{valr-package}
\title{valr: Genome Interval Arithmetic}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read and manipulate genome intervals and signals. Provides functionality similar to command-line tool suites within R, enabling interactive analysis and visualization of genome-scale data. Riemondy et al. (2017) \doi{10.12688/f1000research.11997.1}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rnabioco/valr}
  \item \url{https://rnabioco.github.io/valr/}
  \item Report bugs at \url{https://github.com/rnabioco/valr/issues}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@gmail.com} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})

Authors:
\itemize{
  \item Kent Riemondy \email{kent.riemondy@gmail.com} (\href{https://orcid.org/0000-0003-0750-1273}{ORCID})
}

Other contributors:
\itemize{
  \item RNA Bioscience Initiative (03wmf1y16) [funder, copyright holder]
}

}
\keyword{internal}
