% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valr-package.r
\name{valr}
\alias{valr}
\title{valr: genome interval arithmetic in R}
\description{
valr provides tools to read and manipulate intervals and signals on a genome
reference. valr was developed to facilitate interactive analysis of
genome-scale data sets, leveraging the power of dplyr and piping.
}
\details{
To learn more about valr, start with the vignette:
\code{browseVignettes(package = "valr")}
}
\seealso{
Report bugs at \url{https://github.com/rnabioco/valr/issues}
}
\author{
Jay Hesselberth \href{mailto:jay.hesselberth@gmail.com}{jay.hesselberth@gmail.com}

Kent Riemondy \href{mailto:kent.riemondy@gmail.com}{kent.riemondy@gmail.com}
}
