% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed12_to_exons.r
\name{bed12_to_exons}
\alias{bed12_to_exons}
\title{convert BED12 to exons in BED3}
\usage{
bed12_to_exons(bed12_tbl)
}
\description{
convert BED12 to exons in BED3
}
\examples{
bed12_tbl <- read_bed12('mm9.bed12.gz')
bed3_exon_tbl <- bed12_to_exons(bed12_tbl) 

}

