% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed12_to_exons.r
\name{bed12_to_exons}
\alias{bed12_to_exons}
\title{Convert BED12 to individual exons in BED6.}
\usage{
bed12_to_exons(bed12_tbl)
}
\arguments{
\item{bed12_tbl}{tbl in BED12 format}
}
\description{
After conversion to BED6 format, the \code{score} column contains the exon 
number, with respect to strand (i.e., exon 1 for `-` strand genes will have 
larger start and end coordinates).
}
\examples{
bed12_path <- system.file('extdata', 'mm9.bed12.gz', package = 'Rbedtools')
bed12_tbl <- read_bed12(bed12_path)
bed6_exon_tbl <- bed12_to_exons(bed12_tbl) 

# first exons
subset(bed6_exon_tbl, score == 1)

}

