% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed12_to_exons.r
\name{bed12_to_exons}
\alias{bed12_to_exons}
\title{convert BED12 to individual exons in BED6}
\usage{
bed12_to_exons(bed12_tbl)
}
\arguments{
\item{bed12_tbl}{tbl in BED12 format}
}
\description{
after conversion, the \code{score} column is the exon number, with respect to
strand (i.e., exon 1 for `-` strand genes will have larger coordinates)
}
\examples{
bed12_tbl <- read_bed12('extdata/mm9.bed12.gz')
bed6_exon_tbl <- bed12_to_exons(bed12_tbl) 

# first exons: score == exon number
bed6_exon_tbl \%>\% filter(score == 1)

}

