% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed12_to_exons.r
\name{bed12_to_exons}
\alias{bed12_to_exons}
\title{Convert BED12 to individual exons in BED6.}
\usage{
bed12_to_exons(x)
}
\arguments{
\item{x}{\code{\link[=tbl_interval]{tbl_interval()}}}
}
\description{
After conversion to BED6 format, the \code{score} column contains the exon
number, with respect to strand (i.e., the first exon for \code{-} strand
genes will have larger start and end coordinates).
}
\examples{
x <- read_bed12(valr_example('mm9.refGene.bed.gz'))
bed12_to_exons(x)

}
\seealso{
Other utilities: \code{\link{bed_makewindows}}
}
