% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed12_to_exons.r
\name{bed12_to_exons}
\alias{bed12_to_exons}
\title{Convert BED12 to individual exons in BED6.}
\usage{
bed12_to_exons(x)
}
\arguments{
\item{x}{tbl in BED12 format}
}
\description{
After conversion to BED6 format, the \code{score} column contains the exon 
number, with respect to strand (i.e., the first exon for \code{-} strand
genes will have larger start and end coordinates).
}
\examples{
bed12_path <- valr_example('mm9.bed12.gz')
x <- read_bed12(bed12_path)
bed12_to_exons(x) 

}
\seealso{
Other utils: \code{\link{bed_makewindows}}
}

