% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_closest.r
\name{bed_closest}
\alias{bed_closest}
\title{Identify closest intervals.}
\usage{
bed_closest(x, y, overlap = TRUE, strand = FALSE, strand_opp = FALSE,
  suffix = c(".x", ".y"), distance_type = c("genome", "strand", "abs"))
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{overlap}{include overlapping intervals and report overlap}

\item{suffix}{colname suffixes in output}

\item{distance_type}{reporting method for distance to 
nearest interval (
genome (default) = use negative distances to report upstream intervals,
          strand = define upstream based on strand,
             abs = report absolute value of distance)}
}
\value{
\code{data_frame}
}
\description{
Identify closest intervals.
}
\examples{
x <- tibble::tribble(
  ~chrom, ~start, ~end,
  'chr1',      100,     125
)

y <- tibble::tribble(
  ~chrom, ~start, ~end,
  'chr1',      25,      50,
  'chr1',     140,     175
)
 
bed_glyph(bed_closest(x, y))

x <- tibble::tribble(
~chrom, ~start, ~end,
"chr1", 500,    600,
"chr2", 5000,   6000
) 

y <- tibble::tribble(
~chrom, ~start, ~end,
"chr1", 100,    200,
"chr1", 150,    200,
"chr1", 550,    580,
"chr2", 7000,   8500
) 

bed_closest(x, y)
bed_closest(x, y, overlap = FALSE)


}
\seealso{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/closest.html}

Other multi-set-ops: \code{\link{bed_coverage}},
  \code{\link{bed_intersect}}, \code{\link{bed_map}},
  \code{\link{bed_subtract}}, \code{\link{bed_window}}
}

