% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_cluster.r
\name{bed_cluster}
\alias{bed_cluster}
\title{Cluster neighboring intervals.}
\usage{
bed_cluster(x, max_dist = 0, strand = FALSE)
}
\arguments{
\item{x}{tbl of intervals}

\item{max_dist}{maximum distance between clustered intervals.}

\item{strand}{cluster features on same strand}
}
\value{
\code{data_frame}
}
\description{
Returned \code{data_frame} contains a new \code{.id} column that can be used for
grouping along with \code{chrom}. Default \code{max_dist} of 0 means that
overlapping and book-ended intervals will be clustered.
}
\examples{
x <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", 100,  200,
 "chr1", 180,  250,
 "chr1", 250,  500,
 "chr1", 501,  1000
)

bed_cluster(x)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/cluster.html}
}

