% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bed_complement.r
\name{complement_impl}
\alias{bed_complement}
\alias{complement_impl}
\title{Identify intervals in a genome not covered by a query.}
\usage{
complement_impl(gdf, genome)

bed_complement(x, genome)
}
\arguments{
\item{gdf}{tbl grouped by chrom}

\item{genome}{chrom sizes}

\item{x}{tbl of intervals}

\item{genome}{chrom sizes}
}
\value{
\code{data_frame}
}
\description{
Identify intervals in a genome not covered by a query.
}
\examples{
genome <- tibble::frame_data(
   ~chrom,  ~size,
   "chr1", 500,
   "chr2", 600,
   "chr3", 800
) 

x <- tibble::frame_data(
   ~chrom, ~start, ~end,
   "chr1", 100,    300,
   "chr1", 200,    400,
   "chr2", 1,      100,
   "chr2", 200,    400,
   "chr3", 500,    600
)

# intervals not covered by x
bed_complement(x, genome)

}

