% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_complement.r
\name{bed_complement}
\alias{bed_complement}
\title{identify intervals in a genome that are not covered by a query}
\usage{
bed_complement(bed_tbl, genome)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{genome}{chrom sizes}
}
\value{
\code{data_frame}
}
\description{
identify intervals in a genome that are not covered by a query
}
\examples{
genome <- tibble::frame_data(
   ~chrom,  ~size,
   "chr1", 500,
   "chr2", 600,
   "chr3", 800
) 

bed_tbl <- tibble::frame_data(
   ~chrom, ~start, ~end,
   "chr1", 100,    300,
   "chr1", 200,    400,
   "chr2", 1,      100,
   "chr2", 200,    400,
   "chr3", 500,    600
)

# intervals not covered by bed_tbl
bed_complement(bed_tbl, genome)

}

