% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_coverage.r
\name{bed_coverage}
\alias{bed_coverage}
\title{Compute coverage of intervals.}
\usage{
bed_coverage(x, y, ...)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{...}{extra arguments (not used)}
}
\description{
Compute coverage of intervals.
}
\note{
Book-ended intervals are counted as overlapping by default
}
\examples{
x <- tibble::tribble(
~chrom, ~start, ~end,
"chr1", 100,    500,
"chr2", 200,    400,
"chr2", 300,    500,
"chr2", 800,    900
)

y <- tibble::tribble(
~chrom, ~start, ~end, ~value,
"chr1", 150,    400,  100,
"chr1", 500,    550,  100,
"chr2", 230,    430,  200,
"chr2", 350,    430,  300
)

bed_coverage(x, y)
 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/coverage.html}

Other multi-set-ops: \code{\link{bed_closest}},
  \code{\link{bed_intersect}}, \code{\link{bed_map}},
  \code{\link{bed_subtract}}, \code{\link{bed_window}}
}

