% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_coverage.R
\name{bed_coverage}
\alias{bed_coverage}
\title{Compute coverage of y intervals over x intervals}
\usage{
bed_coverage(x, y, strand = FALSE, strand_opp = FALSE, ...)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{strand}{intersect intervals on same strand}

\item{strand_opp}{intersect intervals on opposite strands}

\item{...}{extra arguments (not used)}
}
\description{
Compute coverage of y intervals over x intervals
}
\note{
Book-ended intervals are counted as overlapping by default
}
\examples{
x <- tibble::tribble(
~chrom, ~start, ~end,
"chr1", 100,    500,
"chr2", 200,    400,
"chr2", 300,    500,
"chr2", 800,    900
)

y <- tibble::tribble(
~chrom, ~start, ~end, ~value,
"chr1", 150,    400,  100,
"chr1", 500,    550,  100,
"chr2", 230,    430,  200,
"chr2", 350,    430,  300
)

bed_coverage(x, y)
 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/coverage.html}
}

