% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_flank.r
\name{bed_flank}
\alias{bed_flank}
\title{create flanking intervals from input intervals}
\usage{
bed_flank(bed_tbl, genome, both = 0, left = 0, right = 0,
  fraction = FALSE, strand = FALSE)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{genome}{tbl of chrom sizes}

\item{both}{number of bases on both sizes}

\item{left}{number of bases on left side}

\item{right}{number of bases on right side}

\item{fraction}{define flanks based on fraction of interval length}

\item{strand}{define \code{left} and \code{right} based on strand}
}
\value{
\code{data_frame}
}
\description{
create flanking intervals from input intervals
}
\examples{
genome <- tibble::frame_data(
 ~chrom, ~size,
 "chr1", 5000
)

bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end, ~name, ~score, ~strand,
 "chr1", 500,    1000, '.',   '.',    '+',
 "chr1", 1000,   1500, '.',   '.',    '-'
)

bed_flank(bed_tbl, left = 100)
bed_flank(bed_tbl, right = 100)
bed_flank(bed_tbl, both = 100)

bed_flank(bed_tbl, both = 0.5, fraction=TRUE)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/flank.html}
}

