% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{bed_glyph}
\alias{bed_glyph}
\title{convienence function for generating example bed glyphs}
\usage{
bed_glyph(x, y = NULL, .fun, ..., x_name = "X", y_name = "Y",
  start_col = "start", end_col = "end", title_ops = "(x, y)")
}
\arguments{
\item{x}{tbl of intervals passed to .fun}

\item{y}{optional tbl of intervals to pass to .fun  (default = FALSE)}

\item{.fun}{valr function to execute on (x and/or y intervals)}

\item{...}{other arguments to pass to .fun (x ivls and y ivls passed by default if supplied)}

\item{x_name}{name of plotted x intervals (default is "X")}

\item{y_name}{name of plotted y intervals (default is "Y")}

\item{start_col}{name of column containing start positions to use for plotting results (default = "start")}

\item{end_col}{name of columm containing end positions to use for plotting results (default = "end")}

\item{title_ops}{options to pass to title display in ggplot object (default = "(x, y"))}
}
\value{
\code{ggplot2_obj}
}
\description{
convienence function for generating example bed glyphs
}

