% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_intersect.r
\name{bed_intersect}
\alias{bed_intersect}
\title{instersect BED intervals}
\usage{
bed_intersect(bed_tbl_a, bed_tbl_b)
}
\arguments{
\item{bed_tbl_a}{BED intervals}

\item{bed_tbl_b}{BED intervals}
}
\description{
instersect BED intervals
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/intersect.html}
}

