% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_intersect.r
\name{bed_intersect}
\alias{bed_intersect}
\title{instersect BED intervals}
\usage{
bed_intersect(df_a, df_b, by_chrom = TRUE)
}
\arguments{
\item{df_a}{BED intervals}

\item{df_b}{BED intervals}
}
\description{
instersect BED intervals
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/intersect.html}
}

