% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/bed_intersect.r
\name{intersect_impl}
\alias{bed_intersect}
\alias{intersect_impl}
\title{Identify intersecting intervals.}
\usage{
intersect_impl(x, y, suffix_x = ".x", suffix_y = ".y")

bed_intersect(x, y, strand = FALSE, strand_opp = FALSE, suffix_x = ".x",
  suffix_y = ".y")
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{suffix_x}{suffix for intersected intervals from x (except chrom)}

\item{suffix_y}{suffix for intersected intervals from y (except chrom)}

\item{strand}{intersect intervals on same strand}

\item{strand_opp}{intersect intervals on opposite strands}
}
\description{
Identify intersecting intervals.
}
\note{
Book-ended intervals have \code{.overlap} values of 0 in the output.
}
\examples{
x <- tibble::frame_data(
~chrom, ~start, ~end,
"chr1", 100,    500,
"chr2", 200,    400,
"chr2", 300,    500,
"chr2", 800,    900
)

y <- tibble::frame_data(
~chrom, ~start, ~end, ~value,
"chr1", 150,    400,  100,
"chr1", 500,    550,  100,
"chr2", 230,    430,  200,
"chr2", 350,    430,  300
)

bed_intersect(x, y)
 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/intersect.html}
}

