% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_intersect.r
\name{bed_intersect}
\alias{bed_intersect}
\title{Identify intersecting intervals.}
\usage{
bed_intersect(x, y, max_dist = 0, strand = FALSE, strand_opp = FALSE)
}
\arguments{
\item{x}{BED intervals}

\item{y}{BED intervals}

\item{max_dist}{maximum distance between intersections}

\item{strand}{intersect intervals on same strand}
}
\description{
Identify intersecting intervals.
}
\examples{
x <- tibble::frame_data(
~chrom, ~start, ~end,
"chr1", 100,    500,
"chr2", 200,    400,
"chr2", 300,    500,
"chr2", 800,    900
)

y <- tibble::frame_data(
~chrom, ~start, ~end,
"chr1", 150,    400,
"chr2", 230,    430,
"chr2", 350,    430
)

bed_intersect(x, y)
bed_intersect(x, y, max_dist = 50)
 
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/intersect.html}
}

