% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\alias{concat}
\alias{distinct}
\alias{distinct_only}
\alias{first}
\alias{last}
\title{Map signals over intervals.}
\usage{
bed_map(bed_tbl, signal_tbl, ...)

concat(.data, sep = ",")

distinct_only(.data, sep = ",")

distinct(.data, sep = ",")

first(.data)

last(.data)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{signal_tbl}{tbl of signals}

\item{...}{name-value pairs of summary functions like \code{\link{min}()},
\code{\link{count}()}, \code{\link{concat}()}}
}
\value{
\code{data_frame}
}
\description{
Map signals over intervals.
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
signal_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)

# mean, median, sd etc
bed_map(bed_tbl, signal_tbl, sum = sum(value))
bed_map(bed_tbl, signal_tbl, min = min(value), max = max(value))

bed_map(bed_tbl, signal_tbl, concat(value))
bed_map(bed_tbl, signal_tbl, distinct(value))
bed_map(bed_tbl, signal_tbl, first(value))
bed_map(bed_tbl, signal_tbl, last(value))

bed_map(bed_tbl, signal_tbl, absmax = abs(max(value)))
bed_map(bed_tbl, signal_tbl, absmin = abs(min(value)))
bed_map(bed_tbl, signal_tbl, count = length(value))
bed_map(bed_tbl, signal_tbl, count_distinct = length(unique(value)))

# use decreasing = TRUE to reverse numbers
bed_map(bed_tbl, signal_tbl, distinct_num = distinct(sort(value)))

}

