% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{absmax}
\alias{absmin}
\alias{bed_map}
\alias{concat}
\alias{count}
\alias{count_distinct}
\alias{distinct}
\alias{first}
\alias{last}
\title{Map signals over intervals}
\usage{
bed_map(bed_tbl, signal_tbl, ...)

absmax(.data)

absmin(.data)

concat(.data, sep = ",")

distinct(.data, sep = ",")

count(.data)

count_distinct(.data)

first(.data)

last(.data)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{signal_tbl}{tbl of signals}

\item{...}{name-value pairs of summary functions like \code{\link{min}()},
\code{\link{count}()}, \code{\link{concat}()}}
}
\value{
\code{data_frame}
}
\description{
Map signals over intervals
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
signal_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)
 
bed_map(bed_tbl, signal_tbl, sum = sum(value), max = max(value))
bed_map(bed_tbl, signal_tbl, concat = concat(value))

}

