% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\alias{concat}
\alias{distinct}
\alias{distinct_only}
\alias{first}
\alias{last}
\title{Map signals over intervals.}
\usage{
bed_map(x, y, ...)

concat(.data, sep = ",")

distinct_only(.data, sep = ",")

distinct(.data, sep = ",")

first(.data)

last(.data)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of signals}

\item{...}{name-value pairs of summary functions like \code{\link{min}()}, 
\code{\link{count}()}, \code{\link{concat}()}. \code{start} and \code{end}
colnames in returned data from .x suffixes.}
}
\value{
\code{data_frame}
}
\description{
Map signals over intervals.
}
\examples{
x <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
y <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)

# mean, median, sd etc
bed_map(x, y, sum = sum(value))
bed_map(x, y, min = min(value), max = max(value))

bed_map(x, y, concat(value))
bed_map(x, y, distinct(value))
bed_map(x, y, first(value))
bed_map(x, y, last(value))

bed_map(x, y, absmax = abs(max(value)))
bed_map(x, y, absmin = abs(min(value)))
bed_map(x, y, count = length(value))
bed_map(x, y, count_distinct = length(unique(value)))

# use decreasing = TRUE to reverse numbers
bed_map(x, y, distinct_num = distinct(sort(value)))

}

