% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\alias{concat}
\alias{distinct}
\alias{distinct_only}
\alias{first}
\alias{last}
\title{Map signals over intervals.}
\usage{
bed_map(x, y, ...)

concat(.data, sep = ",")

distinct_only(.data, sep = ",")

distinct(.data, sep = ",")

first(.data)

last(.data)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of signals}

\item{...}{name-value pairs of summary functions like \code{\link{min}()}, 
\code{\link{count}()}, \code{\link{concat}()}.}

\item{.data}{vector of things}

\item{sep}{separator character}
}
\value{
\code{data_frame}
}
\description{
Map signals over intervals.
}
\note{
Column names have \code{.x} and \code{.y} suffixes.
}
\examples{
x <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
y <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)

# mean, median, sd etc
bed_map(x, y, sum = sum(value.y))
bed_map(x, y, min = min(value.y), max = max(value.y))

bed_map(x, y, concat(value.y))
bed_map(x, y, distinct(value.y))
bed_map(x, y, first(value.y))
bed_map(x, y, last(value.y))

bed_map(x, y, absmax = abs(max(value.y)))
bed_map(x, y, absmin = abs(min(value.y)))
bed_map(x, y, count = length(value.y))
bed_map(x, y, count_distinct = length(unique(value.y)))

# use decreasing = TRUE to reverse numbers
bed_map(x, y, distinct_num = distinct(sort(value.y)))

}

