% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\title{Map signals over intervals}
\usage{
bed_map(interval_tbl, signal_tbl, signal_col, operation = op_choices)
}
\arguments{
\item{interval_tbl}{tbl of intervals}

\item{signal_col}{bare column name in \code{signal_tbl} for \code{operation}}

\item{operation}{operation to perform on intersected intervals. One of mean, median,
sum, min, max, absmin, absmax}

\item{singal_tbl}{tbl of signals}
}
\value{
\code{data.frame}
}
\description{
Map signals over intervals
}
\examples{

signal <- read_bedgraph('inst/extdata/test.bg.gz')
intervals <- read_bed('inst/extdata/3fields.bed.gz')
 
}

