% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\alias{concat}
\alias{first}
\alias{last}
\alias{values}
\alias{values_unique}
\title{Map signals over intervals.}
\usage{
bed_map(x, y, ...)

concat(.data, sep = ",")

values_unique(.data, sep = ",")

values(.data, sep = ",")

first(.data)

last(.data)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of signals}

\item{...}{name-value pairs of summary functions like \code{\link{min}()}, 
\code{\link{count}()}, \code{\link{concat}()}.}

\item{.data}{data}

\item{sep}{separator character}
}
\value{
\code{data_frame}
}
\description{
Multiple mapping functions can be specified.
}
\details{
Variables from the `x` and `y` tables should be used with `.x` and `.y`
suffixes.

Existing grouping variables are maintained in the output, but have `.x` or 
`.y` suffixes`.
}
\note{
Column names have \code{.x} and \code{.y} suffixes.
}
\examples{

x <- tibble::tribble(
~chrom, ~start, ~end,
'chr1',      1,      100
)

y <- tibble::tribble(
  ~chrom, ~start, ~end, ~value,
  'chr1',      1,     20,    10,
  'chr1',      30,    50,    20,
  'chr1',      90,    120,   30
)

bed_glyph(bed_map(x, y, value = sum(value.y)), label = 'value')

x <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
y <- tibble::tribble(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)

# mean, median, sd etc
bed_map(x, y, sum = sum(value.y))
bed_map(x, y, min = min(value.y), max = max(value.y))

bed_map(x, y, concat(value.y))
bed_map(x, y, first(value.y))
bed_map(x, y, last(value.y))

bed_map(x, y, absmax = abs(max(value.y)))
bed_map(x, y, absmin = abs(min(value.y)))
bed_map(x, y, count = length(value.y))
bed_map(x, y, count_distinct = length(unique(value.y)))

bed_map(x, y, vals = values(value.y))
bed_map(x, y, vals.unique = values_unique(value.y))

}
\seealso{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/map.html}
}

