% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_map.r
\name{bed_map}
\alias{bed_map}
\title{Map signals over intervals}
\usage{
bed_map(bed_tbl, signal_tbl, signal_col, operation)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{signal_tbl}{tbl of signals}

\item{signal_col}{bare column name in \code{signal_tbl} for \code{operation}}

\item{operation}{operation to perform on intersected intervals. One of mean, median,
sum, min, max, absmin, absmax}
}
\value{
\code{data_frame}
}
\description{
Map signals over intervals
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 100, 250,
 "chr2", 250, 500)
 
signal_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end, ~value,
 "chr1", 100, 250, 10,
 "chr1", 150, 250, 20,
 "chr2", 250, 500, 500)

bed_map(bed_tbl, signal_tbl, value, sum)

}

