% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_merge.r
\name{bed_merge}
\alias{bed_merge}
\title{Merge overlapping intervals}
\usage{
bed_merge(bed_tbl, max_dist = 0)
}
\arguments{
\item{max_dist}{maximum distance between intervals to merge}
}
\description{
Merge overlapping intervals
}
\note{
TODO: implement strand (group_by?), max_dist
}
\examples{
bed_tbl <- tibble::frame_data(
 ~chrom, ~start, ~end,
 "chr1", 1,      50,
 "chr1", 100,    200,
 "chr1", 150,    250,
 "chr2", 1,      25,
 "chr2", 200,    400,
 "chr2", 400,    500,
 "chr2", 450,    550
)

bed_merge(bed_tbl)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/merge.html}
}

