% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_random.r
\name{bed_random}
\alias{bed_random}
\title{generate randomly placed intervals on a genome}
\usage{
bed_random(genome, length = 100, n = 100, seed = NULL)
}
\arguments{
\item{genome}{genome tbl}

\item{length}{length of intervals}

\item{n}{number of intervals to generate}

\item{seed}{seed RNG for reproducible intervals}
}
\description{
generate randomly placed intervals on a genome
}
\examples{
genome <- dplyr::tibble(
  ~chrom,  ~size,
  "chr1",  10000000,
  "chr2",  50000000,
  "chr3",  60000000,
  "chrX",  5000000
) 

# random intervals (unsorted)
bed_random(genome)

# 500 random intervals of length 500 
bed_random(genome, length = 500, n = 500)

# reproducible random intervals
bed_random(genome, seed = 2016)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/random.html}
}

