% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_shift.r
\name{bed_shift}
\alias{bed_shift}
\title{Adjust intervals by a fixed size.}
\usage{
bed_shift(bed_tbl, genome, size = 0, strand = FALSE, fraction = 0,
  trim = FALSE)
}
\arguments{
\item{bed_tbl}{tbl of intervals}

\item{genome}{chromosome sizes}

\item{size}{number of bases to shift. postive numbers shift right, negative shift left.}

\item{strand}{logical indicating shift according to +/- in strand column}

\item{fraction}{define \code{size} as a fraction of interval}

\item{trim}{adjust coordinates for out-of-bounds intervals}
}
\value{
\code{data.frame}
}
\description{
Out-of-bounds intervals are removed by default.
}
\examples{
bed_tbl <- tibble::frame_data(
   ~chrom, ~start, ~end, ~strand,
   "chr1", 100, 150, "+",
   "chr1", 200, 250, "+",
   "chr2", 300, 350, "+",
   "chr2", 400, 450, "-",
   "chr3", 500, 550, "-",
   "chr3", 600, 650, "-" 
   )

genome <- tibble::frame_data(
   ~chrom, ~size,
   "chr1", 1000,
   "chr2", 2000,
   "chr3", 3000
   )

bed_shift(bed_tbl, genome, 100)
bed_shift(bed_tbl, genome, fraction = 0.5)
bed_shift(bed_tbl, genome, 100, strand = TRUE)
bed_shift(bed_tbl, genome, strand = TRUE, fraction = 0.5)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/shift.html}
}

