% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_slop.r
\name{bed_slop}
\alias{bed_slop}
\title{increase the size of input intervals}
\usage{
bed_slop(bed_df, genome, both = 0, left = 0, right = 0,
  fraction = FALSE, strand = FALSE)
}
\arguments{
\item{genome}{tbl of chrom sizes}

\item{both}{number of bases on both sizes}

\item{left}{number of bases on left side}

\item{right}{number of bases on right side}

\item{fraction}{define flanks based on fraction of interval length}

\item{strand}{define \code{left} and \code{right} based on strand}
}
\value{
\code{data_frame}
}
\description{
increase the size of input intervals
}
\examples{
genome <- tibble::frame_data(
 ~chrom, ~size,
 "chr1", 5000
)

bed_df <- tibble::frame_data(
 ~chrom, ~start, ~end, ~name, ~score, ~strand,
 "chr1", 500,    1000, '.',   '.',     '+',
 "chr1", 1000,   1500, '.',   '.',     '-'
)

bed_df \%>\% bed_slop(genome, left = 100)
bed_df \%>\% bed_slop(genome, right = 100)
bed_df \%>\% bed_slop(genome, both = 100)

bed_df \%>\% bed_slop(genome, both = 0.5, fraction=TRUE)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/slop.html}
}

