% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_sort.r
\name{bed_sort}
\alias{bed_sort}
\title{Sort intervals}
\usage{
bed_sort(intervals, size = FALSE, by_chrom = FALSE, reverse = FALSE)
}
\arguments{
\item{intervals}{tbl of intervals}

\item{size}{sort by interval size}

\item{by_chrom}{sort within chromosome}

\item{reverse}{reverse sort order}
}
\description{
Sort intervals
}
\examples{
bed_df <- dplyr::tibble(
   ~chrom, ~start, ~end,
   "chr8", 500, 1000,
   "chr8", 1000, 5000,
   "chr8", 100, 200,
   "chr1", 100, 300,
   "chr1", 100, 200
)

# sort by chrom and coordinates
bed_df \%>\% bed_sort()

# reverse sort order
bed_df \%>\% bed_sort(reverse = TRUE)

# sort by interval size
bed_df \%>\% bed_sort(size = TRUE)

# sort by interval size within chrom
bed_df \%>\% bed_sort(size = TRUE, by_chrom = TRUE)

}

