% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_subtract.r
\name{bed_subtract}
\alias{bed_subtract}
\title{Subtract intervals.}
\usage{
bed_subtract(x, y, any = FALSE)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{any}{remove any \code{x} intervals that overlap \code{y}}
}
\description{
Subtract \code{y} intervals from \code{x} intervals.
}
\examples{
x <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", 100,    200,
 "chr1", 250,    400,
 "chr1", 500,    600,
 "chr1", 1000,   1200,
 "chr1", 1300,   1500
)

y <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", 150,    175,
 "chr1", 510,    525,
 "chr1", 550,    575,
 # dangling left
 "chr1", 900,    1050,
 # dangling right
 "chr1", 1150,   1250,
 # full containment
 "chr1", 1299,   1501
)

bed_subtract(x, y)
bed_subtract(x, y, any = TRUE)
 
}
\seealso{
\url{http://bedtools.readthedocs.io/en/latest/content/tools/subtract.html}

Other multi-set-ops: \code{\link{bed_closest}},
  \code{\link{bed_coverage}}, \code{\link{bed_intersect}},
  \code{\link{bed_window}}
}

