% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_window.r
\name{bed_window}
\alias{bed_window}
\title{Intersect intervals within a window.}
\usage{
bed_window(x, y, size = 1000)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{size}{size of flank for new intervals}
}
\description{
Intersect intervals within a window.
}
\examples{
x <- tibble::frame_data(
  ~chrom, ~start, ~end,
  "chr1", 100,    200
  )
  
y <- tibble::frame_data(
  ~chrom, ~start, ~end,
  "chr1", 500,    1000,
  "chr1", 1300,   2000
  )

bed_window(x, y)

}

