% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_window.r
\name{bed_window}
\alias{bed_window}
\title{Identify intersecting intervals within a specified distance}
\usage{
bed_window(x, y, genome, ...)
}
\arguments{
\item{x}{tbl of intervals}

\item{y}{tbl of intervals}

\item{genome}{tbl of chrom sizes}

\item{...}{params for bed_slop and bed_intersect}
}
\description{
Identify intersecting intervals within a specified distance
}
\examples{
x <- tibble::frame_data(
  ~chrom, ~start, ~end,
  "chr1", 10,    100,
  "chr2", 200,    400,
  "chr2", 300,    500,
  "chr2", 800,    900
)

y <- tibble::frame_data(
  ~chrom, ~start, ~end,
  "chr1", 150,    400,
  "chr2", 230,    430,
  "chr2", 350,    430
)

genome <- tibble::frame_data(
  ~chrom, ~size,
  "chr1", 500,
  "chr2", 1000
)

bed_window(x, y, genome, both = 100)

}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/content/tools/window.html}
}

