% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome.r
\name{bound_intervals}
\alias{bound_intervals}
\title{select bounded intervals}
\usage{
bound_intervals(bed_tbl, genome)
}
\arguments{
\item{bed_tbl}{a tbl of intervals}

\item{genome}{a tbl of chrom sizes}
}
\description{
select bounded intervals
}

