% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome.r
\name{bound_intervals}
\alias{bound_intervals}
\title{Select intervals bounded by a genome.}
\usage{
bound_intervals(x, genome, trim = FALSE)
}
\arguments{
\item{x}{\link{ivl_df}}

\item{genome}{\link{genome_df}}

\item{trim}{adjust coordinates for out-of-bounds intervals}
}
\value{
\link{ivl_df}
}
\description{
Used to remove out-of-bounds intervals, or trim interval coordinates using a
\code{genome}.
}
\examples{
x <- tibble::tribble(
  ~chrom, ~start, ~end,
  "chr1", -100,   500,
  "chr1", 100,    1e9,
  "chr1", 500,    1000
)

genome <- read_genome(valr_example("hg19.chrom.sizes.gz"))

# out-of-bounds are removed by default ...
bound_intervals(x, genome)

# ... or can be trimmed within the bounds of a genome
bound_intervals(x, genome, trim = TRUE)

}
\seealso{
Other utilities: 
\code{\link{bed12_to_exons}()},
\code{\link{bed_makewindows}()},
\code{\link{flip_strands}()},
\code{\link{interval_spacing}()}
}
\concept{utilities}
