% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strands.r
\name{flip_strands}
\alias{flip_strands}
\title{Flip strands in intervals.}
\usage{
flip_strands(x)
}
\arguments{
\item{x}{\code{\link[=tbl_interval]{tbl_interval()}}}
}
\description{
Flips positive (\code{+}) stranded intervals to negative (\code{-}) strands,
and vice-versa. Facilitates comparisons among intervals on opposing strands.
}
\examples{
x <- trbl_interval(
  ~chrom, ~start, ~end, ~strand,
  'chr1', 1,      100,  '+',
  'chr2', 1,      100,  '-'
)

flip_strands(x)

}
