% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strands.r
\name{flip_strands}
\alias{flip_strands}
\title{Flip strands in intervals.}
\usage{
flip_strands(x)
}
\arguments{
\item{x}{tbl of intervals}
}
\description{
Flips positive (\code{+}) stranded intervals to negative (\code{-}) strands,
and vice-versa. Facilitates comparisons among intervals on opposing strands.
}
\examples{
x <- tibble::tribble(
~chrom, ~start, ~end, ~strand,
'chr1', 1,      100,  '+',
'chr2', 1,      100,  '-' 
)

flip_strands(x)
 
}

