% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbedtools.r
\docType{package}
\name{Rbedtools}
\alias{Rbedtools}
\alias{Rbedtools-package}
\title{Rbedtools: genome arithmetic in R}
\description{
Rbedtools provides tools implemented in the BEDtools suite to 
read and manipulate intervals and signals on a genome reference.
}
\details{
To learn more about Rbedtools, start with the vignette:
\code{browseVignettes(package = "Rbedtools")}
}
\author{
Jay Hesselberth <jay.hesselberth@gmail.com>
}

