% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_bed}
\alias{read_bed}
\alias{read_bed12}
\alias{read_bedgraph}
\alias{read_broadpeak}
\alias{read_narrowpeak}
\title{Read BED and related files.}
\usage{
read_bed(filename, n_fields = 3, col_types = bed12_coltypes, sort = TRUE,
  ...)

read_bed12(filename, ...)

read_bedgraph(filename, ...)

read_narrowpeak(filename, ...)

read_broadpeak(filename, ...)
}
\arguments{
\item{filename}{file or URL}

\item{n_fields}{number fields in the BED file}

\item{col_types}{column type spec for \code{readr::read_tsv}}

\item{sort}{sort the tbl by chrom and start}

\item{...}{options to pass to \code{readr::read_tsv}}
}
\value{
\code{data_frame}
}
\description{
read functions for BED and related formats. Filenames can be
  local file or URLs.
}
\details{
\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}

\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}

\url{https://genome.ucsc.edu/goldenPath/help/bedgraph.html}

\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format12}

\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format13}
}
\examples{
# read_bed assumes 3 field BED format. 
bed3_path <- valr_example('3fields.bed.gz')
bed3_tbl <- read_bed(bed3_path)

bed6_path <- valr_example('6fields.bed.gz')
bed6_tbl <- read_bed(bed6_path, n_fields = 6)

# Result is sorted by chrom and start unless `sort = FALSE`
unsorted_bed_tbl <- read_bed(bed3_path, sort = FALSE)

# use `is_sorted` 
is_sorted(unsorted_bed_tbl)

bed12_path <- valr_example('mm9.bed12.gz')
bed12_tbl <- read_bed12(bed12_path)

bedgraph_path <- valr_example('test.bg.gz')
bedgraph_tbl <- read_bedgraph(bedgraph_path)

narrowpeak_path <- valr_example('sample.narrowPeak.gz')
read_narrowpeak(narrowpeak_path)

broadpeak_path <- valr_example('sample.broadPeak.gz')
read_broadpeak(broadpeak_path)

}

