% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\docType{data}
\name{read_bed}
\alias{bed12_coltypes}
\alias{read_bed}
\alias{read_bed12}
\alias{read_bedgraph}
\title{Read BED and related files.}
\format{An object of class \code{list} of length 12.}
\usage{
read_bed(filename, n_fields = 3, col_types = bed12_coltypes, sort = TRUE,
  factor_cols = TRUE, ...)

read_bed12(filename, ...)

read_bedgraph(filename, ...)

bed12_coltypes
}
\arguments{
\item{filename}{file or URL}

\item{n_fields}{number fields in the BED file}

\item{col_types}{column type spec for \code{readr::read_tsv}}

\item{sort}{sort the tbl by chrom and start}

\item{factor_cols}{factor the \code{chrom} and \code{strand} columns}

\item{...}{options to pass to \code{readr::read_tsv}}
}
\value{
\code{data_frame}
}
\description{
\code{read_bed} reads BED files and \code{read_bedgraph} reads 
  bedGraph files. Return value are \code{data.frame} that is sorted by chrom 
  and start unless otherwise specified.
}
\details{
\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}

\url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}

\url{https://genome.ucsc.edu/goldenPath/help/bedgraph.html}
}
\examples{
# read_bed assumes 3 field BED format. 
bed3_path <- system.file('extdata', '3fields.bed.gz', package = 'valr')
bed3_tbl <- read_bed(bed3_path)

bed6_path <- system.file('extdata', '6fields.bed.gz', package = 'valr')
bed6_tbl <- read_bed(bed6_path, n_fields = 6)

# Result is sorted by chrom and start unless `sort = FALSE`
unsorted_bed_tbl <- read_bed(bed3_path, sort = FALSE)

# use `is_sorted` 
is_sorted(unsorted_bed_tbl)

# chrom and strand are converted to factors unless specified
bed_tbl <- read_bed(bed3_path, factor_cols = FALSE)
is.factor(bed_tbl$chrom)

bed12_path <- system.file('extdata', 'mm9.bed12.gz', package = 'valr')
bed12_tbl <- read_bed12(bed12_path)

bedgraph_path <- system.file('extdata', 'test.bg.gz', package = 'valr')
bedgraph_tbl <- read_bedgraph(bedgraph_path)

}
\seealso{
Other read.data: \code{\link{read_genome}}

Other read.data: \code{\link{read_genome}}

Other read.data: \code{\link{read_genome}}
}
\keyword{datasets}

