% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_bigwig}
\alias{read_bigwig}
\title{Read a bigwig file into a valr compatible tbl}
\usage{
read_bigwig(path, ...)
}
\arguments{
\item{path}{path to bigWig file}

\item{...}{params for \code{cpp11bigwig::read_bigwig()}}
}
\description{
This function will output a 4 column tibble with
zero-based chrom, start, end, value columns.
}
\examples{
read_bigwig(valr_example("hg19.dnase1.bw"))

read_bigwig(valr_example("hg19.dnase1.bw"), as = "GRanges")

}
