% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome.r
\name{read_genome}
\alias{bound_intervals}
\alias{read_genome}
\title{read genome files (i.e., UCSC "chromSize" files)}
\usage{
read_genome(filename)

bound_intervals(x, genome, trim = FALSE)
}
\arguments{
\item{filename}{file containing chrom/contig names and sizes,
one-pair-per-line, tab-delimited}

\item{x}{a tbl of intervals}

\item{genome}{a tbl of chrom sizes}

\item{trim}{adjust coordinates for out-of-bounds intervals}
}
\value{
\code{data_frame} with colnames \code{chrom} and \code{size}, sorted
  by \code{size}

\code{data_frame}
}
\description{
Genome files contain chromosome name and size information. These sizes are 
used by downstream functions to identify computed intervals that have ends
outside the defined size.

Select intervals bounded by a genome.
}
\examples{
genome <- valr_example('hg19.chrom.sizes.gz')
read_genome(genome)

\dontrun{
# can also read from a URL
read_genome('https://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes')
}

x <- tibble::tribble(
 ~chrom, ~start, ~end,
 "chr1", -100,   500,
 "chr1", 100,    1e9,
 "chr1", 500,    1000
)

genome <- read_genome(valr_example('hg19.chrom.sizes.gz'))

bound_intervals(x, genome)
bound_intervals(x, genome, trim = TRUE)

}

