% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome.r
\name{read_genome}
\alias{read_genome}
\title{read genome files (i.e., UCSC "chromSize" files)}
\usage{
read_genome(filename)
}
\arguments{
\item{filename}{file containing chrom/contig names and sizes,
one-pair-per-line, tab-delimited}
}
\value{
\code{data_frame} with colnames \code{chrom} and \code{size}, sorted
  by \code{size}
}
\description{
Genome files contain chromosome name and size information. These sizes are 
used by downstream functions to identify computed intervals that have ends
outside the defined size.
}
\examples{
genome <- system.file('extdata', 'hg19.chrom.sizes.gz', package = 'Rbedtools')
read_genome(genome)
read_genome('https://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes')

}

