% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.r
\name{read_gtf}
\alias{read_gtf}
\title{Import and convert GTF/GFF into valr compatible bed tbl format}
\usage{
read_gtf(path, zero_based = TRUE)
}
\arguments{
\item{path}{path to gtf or gff file}

\item{zero_based}{if TRUE, convert to zero based}
}
\description{
This function will output a tibble with
required chrom, start, and end, as well as other columns depending
on content in GTF/GFF file.
}
\examples{

gtf <- read_gtf(valr_example('hg19.gencode.gtf.gz'))
head(gtf)

}
