% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-cpp11bigwig.r, R/reexport-tibble.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read_bigwig}
\alias{read_bigbed}
\alias{data_frame}
\alias{as_data_frame}
\alias{tribble}
\alias{tibble}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cpp11bigwig}{\code{\link[cpp11bigwig]{read_bigbed}}, \code{\link[cpp11bigwig]{read_bigwig}}}

  \item{tibble}{\code{\link[tibble:deprecated]{as_data_frame}}, \code{\link[tibble]{as_tibble}}, \code{\link[tibble:deprecated]{data_frame}}, \code{\link[tibble]{tibble}}, \code{\link[tibble]{tribble}}}
}}

