% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{shared_groups}
\alias{shared_groups}
\title{Compare two tbl_dfs and find the minimum shared groups}
\usage{
shared_groups(x, y)
}
\arguments{
\item{x}{tbl}

\item{y}{tbl}
}
\value{
\code{list}
}
\description{
\code{shared_groups} returns the minimum shared groups between
\code{x} and \code{y} tbls. Return value is \code{NULL} if there are
no shared groups.
}

