% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{tbl_interval}
\alias{tbl_interval}
\alias{tbl_ivl}
\title{Tibble for intervals.}
\usage{
tbl_interval(..., validate = TRUE)

tbl_ivl(..., validate = TRUE)
}
\arguments{
\item{...}{params for \code{\link[tibble]{tibble}}}

\item{validate}{check valid column names}
}
\description{
Inherits from \code{\link[tibble]{tibble}} and enforces \code{chrom},
\code{start} and \code{end} columns.
}
