% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{tbl_sizes}
\alias{tbl_sizes}
\title{create a sizes tibble}
\usage{
tbl_sizes(...)
}
\arguments{
\item{...}{A set of name-value pairs. Arguments are evaluated sequentially,
so you can refer to previously created variables.}
}
\description{
Inherits from \code{[tibble](tibble)} and enforces \code{chrom} and
\code{size} columns. Identical to UCSC "chromSizes" files.
}
