% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbls.r
\name{tbl_sizes}
\alias{tbl_sizes}
\alias{tbl_szs}
\title{Tibble for reference sizes.}
\usage{
tbl_sizes(..., validate = TRUE)

tbl_szs(..., validate = TRUE)
}
\arguments{
\item{...}{params for \code{\link[tibble]{tibble}}}

\item{validate}{check valid column names}
}
\description{
Inherits from \code{\link[tibble]{tibble}} and enforces \code{chrom} and
\code{size} columns. Identical to UCSC "chromSizes" files.
}
