% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valr.r
\docType{package}
\name{valr}
\alias{valr}
\alias{valr-package}
\alias{valr-package}
\title{valr: genome interval arithmetic in R}
\description{
valr provides tools to read and manipulate intervals and signals on a genome
reference. valr was developed to facilitate interactive analysis of
genome-scale data sets, leveraging the power of dplyr and piping.
}
\details{
To learn more about valr, start with the vignette:
\code{browseVignettes(package = "valr")}
}
\seealso{
\url{http://bedtools.readthedocs.org/en/latest/index.html}

\url{https://pythonhosted.org/pybedtools/}

\url{http://bedops.readthedocs.org/en/latest/index.html}

\url{https://bioconductor.org/packages/release/bioc/html/GenomicRanges.html}

\url{https://CRAN.R-project.org/package=bedr}
}
\author{
Jay Hesselberth \href{mailto:jay.hesselberth@gmail.com}{jay.hesselberth@gmail.com}

Kent Riemondy \href{mailto:kent.riemondy@gmail.com}{kent.riemondy@gmail.com}
}
