% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{valr_example}
\alias{valr_example}
\title{Provide working directory for valr example files.}
\usage{
valr_example(path)
}
\arguments{
\item{path}{path to file}
}
\description{
Provide working directory for valr example files.
}
\examples{
valr_example('hg19.chrom.sizes.gz')

}
